requirements.txt
clarifai==9.10.4
streamlit==1.24.0
databricks-connect==13.3
stqdm==0.0.5